new blockchain.py file
the blockchain.py files is responcible for hashing and creating blocks, as well as resolving conclicts and requesting peer chains. This version of the file is not functional and still needs to be integrated into the main routes.py file to allow functioning API calls against the blockchain code in this file.